Removal of unnecessary mallocs and correction of input array
Leaving only the essential line malloc for reading of the data file. Also, the column number of the input array was incorrect, hindering the learning process with less data parameters. Now the training process completes fully and all the data parameters are seen, leading to even better results than before.