Correcting Huber Loss formula and turning off random row selection
Huber Loss formula had an incorrectly squared delta, now corrected.
Also, random row selection for training interferred with the error function too much, so it is deactivated. The shuffling of data is, for now, up to the user. The error function values are now what they should have always been, marking a milestone of sorts in this project.